requirements.txt
fastapi
uvicorn
pydantic
pydantic-settings

sqlmodel
sqlalchemy
alembic
asyncpg
pandas
google-cloud-bigquery
